Adjust File store to new structure
